Started fixing hopgui. Now drawing timeline properly
