+++ tests/CMakeLists.txt
@@ -8,6 +8,8 @@ else()
 endif()
 
 add_executable (Tsc_test Tsc_test.cpp ${platform_src} ${ROOT_DIR}/common/Utils.cpp)
+target_compile_definitions( Tsc_test PUBLIC HOP_ENABLED )
+
 add_executable (Pid_test Pid_test.cpp ${platform_src} )
 
 add_test (NAME TscTest COMMAND Tsc_test)
